+++ build.rs
@@ -10,6 +10,7 @@ fn main() {
 
     println!("cargo:rerun-if-changed=ui/src");
     println!("cargo:rerun-if-changed=ui/package.json");
+    println!("cargo:rerun-if-changed=ui/package-lock.json");
     println!("cargo:rerun-if-changed=ui/svelte.config.js");
     println!("cargo:rerun-if-changed=ui/vite.config.js");